move files to src, update README.md
